pyproject.toml
[project]
name = "city-explorer-agent"
version = "0.1.0"
description = "An educational AI agent that teaches the difference between Tools and Skills"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]
